+++ CHANGELOG.md
@@ -1,10 +1,17 @@
+## [Unreleased] - 2024-02-03
+
+### Added
+
+- Implemented first-in-first-out Queue, and first-in-last-out Stack.
+- Added test cases for Stack and Queue.
+
 ## [Unreleased] - 2024-01-24
 
 ### Added
 
 - Added BaseComparer and DefaultComprarer for comparison logic within collections.
 - Implemented doubly LinkedList, and LinkedListNode.
-- 
+- Added test cases for DefaultComparer and LinkedList.
 
 ## [Unreleased] - 2024-01-21